requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.9.0
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==4.5.2
APScheduler==3.10.4
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
deep-translator
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
praw==7.8.1
prawcore==2.4.0
propcache==0.2.0
pygame==2.6.1
python-decouple==3.8
python-dotenv==1.0.0
python-telegram-bot==21.6
pytz==2025.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
yarl==1.15.2
